Refactor carousel builder into helper modules
