+++ src/helpers/carousel/index.js
@@ -1,3 +1,6 @@
 export * from "./scroll.js";
 export * from "./navigation.js";
 export * from "./accessibility.js";
+export * from "./responsive.js";
+export * from "./cards.js";
+export * from "./focus.js";
